Music Playlist Program .c
The provided program is a music playlist manager that allows users to create a cyclic playlist of songs. The program prompts the user to enter the number of songs and their titles, then adds them to the playlist in sequential order. The playlist is designed to be cyclic, meaning that once it reaches the end, it loops back to the beginning. The program also displays the playlist to the user.